requirements.txt
# DBDiff-Exporter
# 用于比较和监控不同数据库之间数据一致性的 Prometheus 导出器

# Web 框架和服务器
fastapi==0.109.2
uvicorn[standard]==0.27.1

# 监控和指标
prometheus-client==0.19.0

# 数据库驱动
oracledb==2.0.0
psycopg2-binary==2.9.9

# 配置和数据处理
pyyaml==6.0.1
python-multipart==0.0.9
pydantic==2.6.1
typing-extensions==4.9.0 